Alt+Enter is not working a mode switcher, changed to Alt+P
It looks like the key to switch modes cannot be related to Enter
because it gives some problems (sometimes we will press Alt+Enter by
mistake, because we are pressing the Alt modifier for the previous key
combination).

We'll give a try to Alt+P at the moment, which is similar on
QWERTY (we have to use the right-hand pinky) and it's not in the most
important key (Enter).